Lint and some refactoring
