+++ app/src/main/java/me/vanpetegem/accentor/data/tracks/Track.kt
@@ -1,6 +1,8 @@
 package me.vanpetegem.accentor.data.tracks
 
+import android.util.SparseArray
 import java.time.Instant
+import me.vanpetegem.accentor.data.albums.Album
 
 data class Track(
     val id: Int,
@@ -19,6 +21,17 @@ data class Track(
     val locationId: Int?
 ) {
     fun stringifyTrackArtists() = trackArtists.sortedBy { ta -> ta.order }.joinToString(" / ") { ta -> ta.name }
+
+    fun compareTo(other: Track, albums: SparseArray<Album>): Int {
+        val a1 = albums[this.albumId]
+        val a2 = albums[other.albumId]
+        if (a1 == null && a2 == null) { return this.number - other.number }
+        if (a1 == null) { return 1 }
+        if (a2 == null) { return -1 }
+        val order = a1.compareToByName(a2)
+        if (order != 0) { return order }
+        return this.number - other.number
+    }
 
     companion object {
         const val ALBUMARTIST = "me.vanpetegem.accentor.data.tracks.Track.ALBUMARTIST"
